Add files via upload
This is a "b" version as the changes made here are not carried through to the subsequent examples. This version uses key values set directly within the map using keydef + keyword.  The problem with this approach is that it limits reuse of the key definitions because you have to define them in each map. So while this version uses keys, it is a tangent from the way that keys are implemented in the subsequent versions of this document.